Move magic system-related patches to a separate file
